+++ apps/mobile/screens/app-screens.tsx
@@ -16,7 +16,8 @@ import { uploadLocalFileWithPresigned } from '@/lib/file-storage-api';
 import { saveRemoteMediaToLibrary, shareMediaLink } from '@/lib/media-download';
 import { updateProfileImage, updateUsername } from '@/lib/user-api';
 import { getMediaDownloadUrl } from '@/lib/user-media-api';
-import { useHarucutStore } from '@/store/use-harucut-store';
+import { useLibraryStore } from '@/store/use-library-store';
+import { useSessionStore } from '@/store/use-session-store';
 
 type HarucutThemeColors = ReturnType<typeof useHarucutTheme>['colors'];
 
@@ -126,14 +127,14 @@ export function HomeScreen() {
   const { colors, styles } = useAppScreenTheme();
   const router = useRouter();
   const push = (path: string) => router.push(path as never);
-  const historyItems = useHarucutStore((state) => state.historyItems);
-  const historyError = useHarucutStore((state) => state.historyError);
-  const historyStatus = useHarucutStore((state) => state.historyStatus);
-  const loadRemoteFrames = useHarucutStore((state) => state.loadRemoteFrames);
-  const accessMode = useHarucutStore((state) => state.accessMode);
-  const loadRemoteHistory = useHarucutStore((state) => state.loadRemoteHistory);
-  const savedFrames = useHarucutStore((state) => state.savedFrames);
-  const user = useHarucutStore((state) => state.user);
+  const historyItems = useLibraryStore((state) => state.historyItems);
+  const historyError = useLibraryStore((state) => state.historyError);
+  const historyStatus = useLibraryStore((state) => state.historyStatus);
+  const loadRemoteFrames = useLibraryStore((state) => state.loadRemoteFrames);
+  const accessMode = useSessionStore((state) => state.accessMode);
+  const loadRemoteHistory = useLibraryStore((state) => state.loadRemoteHistory);
+  const savedFrames = useLibraryStore((state) => state.savedFrames);
+  const user = useSessionStore((state) => state.user);
 
   const todayMoment = useCurrentDateLabel();
   const recentItems = historyItems.slice(0, 4);
@@ -291,13 +292,13 @@ export function HistoryScreen() {
   const { colors, styles } = useAppScreenTheme();
   const router = useRouter();
   const push = (path: string) => router.push(path as never);
-  const accessMode = useHarucutStore((state) => state.accessMode);
-  const historyError = useHarucutStore((state) => state.historyError);
-  const historyItems = useHarucutStore((state) => state.historyItems);
-  const historyStatus = useHarucutStore((state) => state.historyStatus);
-  const loadRemoteHistory = useHarucutStore((state) => state.loadRemoteHistory);
-  const renameHistoryItem = useHarucutStore((state) => state.renameHistoryItem);
-  const showNotice = useHarucutStore((state) => state.showNotice);
+  const accessMode = useSessionStore((state) => state.accessMode);
+  const historyError = useLibraryStore((state) => state.historyError);
+  const historyItems = useLibraryStore((state) => state.historyItems);
+  const historyStatus = useLibraryStore((state) => state.historyStatus);
+  const loadRemoteHistory = useLibraryStore((state) => state.loadRemoteHistory);
+  const renameHistoryItem = useLibraryStore((state) => state.renameHistoryItem);
+  const showNotice = useSessionStore((state) => state.showNotice);
   const [filter, setFilter] = useState<'ALL' | 'PHOTO' | 'VIDEO'>('ALL');
   const [search, setSearch] = useState('');
   const [editingId, setEditingId] = useState<string | null>(null);
@@ -507,13 +508,13 @@ export function MyPageScreen() {
   const { colors, isDark, styles } = useAppScreenTheme();
   const router = useRouter();
   const replace = (path: string) => router.replace(path as never);
-  const user = useHarucutStore((state) => state.user);
-  const refreshUserProfile = useHarucutStore((state) => state.refreshUserProfile);
-  const setUserProfile = useHarucutStore((state) => state.setUserProfile);
-  const enterAnonymousMode = useHarucutStore((state) => state.enterAnonymousMode);
-  const showNotice = useHarucutStore((state) => state.showNotice);
-  const themePreference = useHarucutStore((state) => state.themePreference);
-  const setThemePreference = useHarucutStore((state) => state.setThemePreference);
+  const user = useSessionStore((state) => state.user);
+  const refreshUserProfile = useSessionStore((state) => state.refreshUserProfile);
+  const setUserProfile = useSessionStore((state) => state.setUserProfile);
+  const enterAnonymousMode = useSessionStore((state) => state.enterAnonymousMode);
+  const showNotice = useSessionStore((state) => state.showNotice);
+  const themePreference = useSessionStore((state) => state.themePreference);
+  const setThemePreference = useSessionStore((state) => state.setThemePreference);
   const [username, setUsername] = useState(user.username);
   const [oldPassword, setOldPassword] = useState('');
   const [newPassword, setNewPassword] = useState('');
